refactor: MinCostPathTailFilter down to costfilter package
